Add support for a separate cache directory
Storing the cached images together with the original images has two
issues

The first issue is when the application can read the directory structure
containing the images, but cannot write to it.

The second issue is that there might be file name collisions; for
example, I acutally have a folder named cache in my picture folder
structure.

Being able to store the cache file in a separate folder has also the
advantage that it makes easier to delete the cache; just delete the
cache folder.
There is no risk to delete an image by accident